refactor: improve code organization and reduce duplication
- Create timeutil package for time-related functions

- Move Tuple type to libs/types.go

- Update files to use common types and functions

- Remove duplicated code across files
