+++ commands/prepare.go
@@ -3,86 +3,13 @@ package commands
 import (
 	"matchmaker/libs"
 	"matchmaker/libs/gcalendar"
+	"matchmaker/libs/timeutil"
 	"matchmaker/util"
 	"os"
 	"path/filepath"
-	"time"
 
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
-
-func FirstDayOfISOWeek(weekShift int) time.Time {
-	date := time.Now()
-	date = time.Date(date.Year(), date.Month(), date.Day(), date.Hour(), 0, 0, 0, date.Location())
-	date = date.AddDate(0, 0, 7*weekShift)
-
-	// iterate to Monday
-	for !(date.Weekday() == time.Monday && date.Hour() == 0) {
-		date = date.Add(time.Hour)
-	}
-
-	date = date.AddDate(0, 0, 0)
-
-	return date
-}
-
-func GetWorkRange(beginOfWeek time.Time, day int, startHour int, startMinute int, endHour int, endMinute int) *libs.Range {
-	start := time.Date(
-		beginOfWeek.Year(),
-		beginOfWeek.Month(),
-		beginOfWeek.Day()+day,
-		startHour,
-		startMinute,
-		0,
-		0,
-		beginOfWeek.Location(),
-	)
-	end := time.Date(
-		beginOfWeek.Year(),
-		beginOfWeek.Month(),
-		beginOfWeek.Day()+day,
-		endHour,
-		endMinute,
-		0,
-		0,
-		beginOfWeek.Location(),
-	)
-	return &libs.Range{
-		Start: start,
-		End:   end,
-	}
-}
-
-func GetWeekWorkRanges(beginOfWeek time.Time) chan *libs.Range {
-	ranges := make(chan *libs.Range)
-
-	go func() {
-		for day := 0; day < 5; day++ {
-			ranges <- GetWorkRange(beginOfWeek, day,
-				viper.GetInt("workingHours.morning.start.hour"),
-				viper.GetInt("workingHours.morning.start.minute"),
-				viper.GetInt("workingHours.morning.end.hour"),
-				viper.GetInt("workingHours.morning.end.minute"))
-			ranges <- GetWorkRange(beginOfWeek, day,
-				viper.GetInt("workingHours.afternoon.start.hour"),
-				viper.GetInt("workingHours.afternoon.start.minute"),
-				viper.GetInt("workingHours.afternoon.end.hour"),
-				viper.GetInt("workingHours.afternoon.end.minute"))
-		}
-		close(ranges)
-	}()
-
-	return ranges
-}
-
-func ToSlice(c chan *libs.Range) []*libs.Range {
-	s := make([]*libs.Range, 0)
-	for r := range c {
-		s = append(s, r)
-	}
-	return s
-}
 
 func loadProblem(weekShift int, groupFile string) *libs.Problem {
 	groupPath := filepath.Join("groups", groupFile)
@@ -97,12 +24,12 @@ func loadProblem(weekShift int, groupFile string) *libs.Problem {
 	util.PanicOnError(err, "Can't get Google Calendar client")
 	util.LogInfo("Connected to google calendar", nil)
 
-	beginOfWeek := FirstDayOfISOWeek(weekShift)
+	beginOfWeek := timeutil.FirstDayOfISOWeek(weekShift)
 	util.LogInfo("Planning for week", map[string]interface{}{
 		"weekFirstDay": beginOfWeek,
 	})
 
-	workRanges := ToSlice(GetWeekWorkRanges(beginOfWeek))
+	workRanges := timeutil.ToSlice(timeutil.GetWeekWorkRanges(beginOfWeek))
 	busyTimes := []*libs.BusyTime{}
 	for _, person := range people {
 		if person.MaxSessionsPerWeek == 0 {
